Created pages directory to separate top level pages from components
