+++ README.md
@@ -45,6 +45,22 @@ npm start
 
 This will start a webserver for you at `http://localhost:8008`, and run a `watch` process so that your front-end assets will be regenerated as you make changes.
 
+### Directory and naming conventions
+
+JS
+* `lib/`: Non-react modules, as well as entry-point modules like `main.jsx` and `routes.jsx`
+* `components/`: Re-usable react components that can be used throughout the site
+* `pages/`: React "page" components, i.e. top-level pages required by `lib/router`
+* `mixins/`: React mixins
+
+Less
+* `less/`: This is where you should add styles. `common.less` is the entry-point.
+
+Other
+* `test/`: For js tests, manual tests
+* `img/`: For images
+* `dist/`: Generated site assets goes here. **This folder is gitignored, do not edit files here.**
+
 ### Test
 
 Fully testing the code is accomplished by running `npm test`,
