+++ pages/clubs.jsx
@@ -1,14 +1,14 @@
 var _ = require('underscore');
 var React = require('react/addons');
 
-var Page = require('./page.jsx');
-var HeroUnit = require('./hero-unit.jsx');
-var Map = require('./map.jsx');
-var Blockquote = require('./blockquote.jsx');
-var IconLinks = require('./icon-links.jsx');
-var IconLink = require('./icon-link.jsx');
-var PageEndCTA = require('./page-end-cta.jsx');
-var Modal = require('./modal.jsx');
+var Page = require('../components/page.jsx');
+var HeroUnit = require('../components/hero-unit.jsx');
+var Map = require('../components/map.jsx');
+var Blockquote = require('../components/blockquote.jsx');
+var IconLinks = require('../components/icon-links.jsx');
+var IconLink = require('../components/icon-link.jsx');
+var PageEndCTA = require('../components/page-end-cta.jsx');
+var Modal = require('../components/modal.jsx');
 var ModalManagerMixin = require('../mixins/modal-manager');
 var TeachAPIClientMixin = require('../mixins/teach-api-client');
