+++ test/browser/clubs-page.test.jsx
@@ -4,7 +4,7 @@ var TestUtils = React.addons.TestUtils;
 
 var StubTeachAPI = require('./stub-teach-api');
 var stubContext = require('./stub-context.jsx');
-var ClubsPage = require('../../components/clubs-page.jsx');
+var ClubsPage = require('../../pages/clubs.jsx');
 
 describe("ClubsPage", function() {
   var clubsPage, teachAPI, xhr;
